// src/app/model/hygiene.model.ts
export interface Hygiene {
    id: number;
    titre: string;
    contenu: string;
    vue: boolean;
}
//
export interface Information {
    id: number;
    titre: string;
    contenu: string;
    cause: string;
    manifestation: string;
    type: string;
    vue: boolean;
}

export interface Soin {
    id: number;
    titre: string;
    contenu: string;
}